sort out reward dtos

+++ backend/src/main/java/com/codecool/tasx/controller/dto/reward/RewardUpdateRequestDto.java
@@ -1,4 +1,4 @@
 package com.codecool.tasx.controller.dto.reward;
 
-public record RewardUpdateRequestDto(Long userId, Long rewardId, String name, String description, int pointCost) {
+public record RewardUpdateRequestDto(Long userId, Long rewardId, String name, String description, long pointCost) {
 }
